requirements.txt
mcp-ai
openai
python-dotenv
aiohttp
asyncio
pydantic